Rename widths to something more specific
